+++ CHANGELOG.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added rollup-exception handling to ioctl-echo-loop
+- Added --reject-inspects as an option to ioctl-echo-loop
+
 ### Changed
-- Lock dependencies versions using Cargo.lock on http-dispatcher and echo-dapp
-- Fix --reject option on ioctl-echo-loop
-- Add --reject-inspects as an option to ioctl-echo-loop
+- Enabled rollup-exception in rollup
+- Updated yield to the new yield\_request format
+- Removed --ack from yield
 
 ## [0.5.1] - 2022-02-22
 ### Changed